Fixed Bugs

+++ app/src/main/java/com/zillennium/utswap/module/finance/lockUpScreen/adapter/LockUpAdapter.kt
@@ -37,7 +37,7 @@ class LockUpAdapter(var arrayList: ArrayList<Any>, var onClickAdapter: OnClickAd
         val buyBackList: LockUpBalanceObject.LockUpBalanceItem = arrayList[position] as LockUpBalanceObject.LockUpBalanceItem
 
         holder.txtTitle.text = buyBackList.name
-        holder.txtAmount.text = "$" + groupingSeparator(buyBackList.action_balance)
+        holder.txtAmount.text = "$" + groupingSeparator(buyBackList.action_balance.toDouble())
         holder.txtDateStart.text = buyBackList.addtimeReable
         holder.txtDateEnd.text = buyBackList.endtimeReable
         holder.txtDuration.text = "${buyBackList.lock_period} Day(s) left"
